backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
